# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Manager API",
        "description": "API Documentation for the Project Management System. <br> **Base URL**: /api",
        "contact": {
            "email": "admin@company.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/assignments/project-teams": {
            "post": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Assign Teams to a Project",
                "description": "Replaces the list of teams assigned to a project. Requires Executive role.",
                "operationId": "assignTeamsToProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "team_ids"
                                ],
                                "properties": {
                                    "project_id": {
                                        "description": "The ID of the project to update",
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "team_ids": {
                                        "description": "Array of Team IDs to assign. Existing assignments not in this list will be removed.",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assignments updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Teams assigned successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden. Only Executives can assign teams."
                    },
                    "422": {
                        "description": "Validation Error (e.g. Invalid Project ID or Team ID)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/assignments/advisors": {
            "post": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Assign Internal Advisor",
                "description": "Grants a user specific 'Contextual Access' to a project, bypassing team restrictions.",
                "operationId": "assignAdvisor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "user_id": {
                                        "description": "The Manager/Associate to add as an advisor",
                                        "type": "integer",
                                        "example": 102
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Advisor assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Advisor assigned successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Business Rule Exception (e.g. User is inactive)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/assignments/advisors/{project}/{userId}": {
            "delete": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Remove Internal Advisor",
                "description": "Revokes the contextual access for a specific user on a specific project.",
                "operationId": "removeAdvisor",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID of the advisor to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Advisor removed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Advisor removed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}